requirements.txt
fastapi
uvicorn
python-multipart
pillow
torch
torchvision
timm
numpy 